[Refactor] Split lib.py into 3 modules in 1 subpackage
