+++ src/lmd/lib/__init__.py
@@ -1,0 +1,4 @@
+"""Core logic of the library"""
+
+from ._geom import Collection, Shape
+from ._segmentation import SegmentationLoader
